Refactored chat components into managable modules
